fix: narrow automation renderer contracts
